dius: use platform configuration for controlling some ifdefs
