+++ include/platform/linux/dius/config.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define DIUS_HAVE_CLOCK_NANOSLEEP
+#define DIUS_USE_MKNOD_TO_CREATE_FILES
